Rewrite controlling logic to fix deadlock and reloading ineffective
When two config reload request is sent in a very short duration,
the second one will be sometimes ignored and may even cause a deadlock
when the first reload is not completed.
This fixes this problem by changing the reload condition from a signal flag
to actually differentiating the config. Also, don't check for config
changes until restart is completed.